godamlah_hackathon/requirement.txt
pip freeze > requirements.txt

supabase==0.0.25
python-dotenv==1.0.0
